DOP-5414: Moving page rating from action bar to on this page (#1383)
